RHOAIENG-79880: Extract getStoredPreferredProject utility to @odh-dashboard/ui-core
