+++ frontend/src/pages/modelServing/screens/global/ModelServingGlobal.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { useNavigate } from 'react-router';
 import TitleWithIcon from '@odh-dashboard/ui-core/design/TitleWithIcon';
 import { ApplicationsPage } from '@odh-dashboard/ui-core';
+import { getStoredPreferredProject } from '@odh-dashboard/ui-core/context/getStoredPreferredProject';
 import { ModelServingContext } from '#~/pages/modelServing/ModelServingContext';
 import useServingPlatformStatuses from '#~/pages/modelServing/useServingPlatformStatuses';
 import { getProjectModelServingPlatform } from '#~/pages/modelServing/screens/projects/utils';
 import { ProjectObjectType } from '#~/concepts/design/utils';
-import { getStoredPreferredProject } from '#~/concepts/projects/getStoredPreferredProject';
 import EmptyModelServing from './EmptyModelServing';
 import InferenceServiceListView from './InferenceServiceListView';
 import ModelServingProjectSelection from './ModelServingProjectSelection';
