+++ frontend/src/pages/pipelines/global/GlobalPipelineCoreLoader.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 import { Navigate, Outlet, useParams } from 'react-router-dom';
 import { ApplicationsPage } from '@odh-dashboard/ui-core';
+import { getStoredPreferredProject } from '@odh-dashboard/ui-core/context/getStoredPreferredProject';
 import { byName, ProjectsContext } from '#~/concepts/projects/ProjectsContext';
 import PipelineCoreNoProjects from '#~/pages/pipelines/global/PipelineCoreNoProjects';
 import { PipelineContextProvider } from '#~/concepts/pipelines/context';
 import InvalidProject from '#~/concepts/projects/InvalidProject';
 import { ProjectObjectType } from '#~/concepts/design/utils';
-import { getStoredPreferredProject } from '#~/concepts/projects/getStoredPreferredProject';
 import PipelineCoreProjectSelector from './PipelineCoreProjectSelector';
 
 type ApplicationPageProps = React.ComponentProps<typeof ApplicationsPage>;
