+++ packages/ui-core/package.json
@@ -61,6 +61,7 @@
     "./context/LocalQueuesContext": "./src/context/LocalQueuesContext.tsx",
     "./context/ProjectHardwareProfilesContext": "./src/context/ProjectHardwareProfilesContext.tsx",
     "./context/ProjectsContext": "./src/context/ProjectsContext.tsx",
+    "./context/getStoredPreferredProject": "./src/context/getStoredPreferredProject.ts",
     "./contexts/ThemeContext": "./src/contexts/ThemeContext.tsx",
     "./contexts/AnalyticsContext": "./src/contexts/AnalyticsContext.tsx",
     "./contexts/NotificationContext": "./src/contexts/NotificationContext.tsx",
